requirement.txt
openai-whisper
ffmpeg
requests
nltk